# api/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "G.D.U.L Sampath - Weather Data Capture",
    "description": "Weather Data Capture For Recherch MSc IT SLIIT",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3010",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "WeatherData",
      "description": "API for Weather Data in the system"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/weather/saveWeatherData": {
      "post": {
        "tags": [
          "WeatherData"
        ],
        "description": "Create new weather data in system",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Weather data that we want to save",
            "schema": {
              "$ref": "#/definitions/capturedWeatherData"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Weather data is successfully saved",
            "schema": {
              "$ref": "#/definitions/capturedWeatherData"
            }
          }
        }
      }
    },

    "/user/saveUser": {
      "post": {
        "tags": [
          "UserData"
        ],
        "description": "Create new user in system",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User data that we want to save",
            "schema": {
              "$ref": "#/definitions/userData"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User data is successfully saved",
            "schema": {
              "$ref": "#/definitions/userData"
            }
          }
        }
      }
    },

    "/sensor/saveSensor": {
      "post": {
        "tags": [
          "SensorData"
        ],
        "description": "Create new Sensor in system",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Sensor data that we want to save",
            "schema": {
              "$ref": "#/definitions/sensorData"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Sensor data is successfully saved",
            "schema": {
              "$ref": "#/definitions/sensorData"
            }
          }
        }
      }
    },

    "/sensor/getStationSensorData": {
      "post": {
        "tags": [
          "SensorData"
        ],
        "description": "Get Station Sensor Data",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Get Station Sensor Data that we want to save",
            "schema": {
              "$ref": "#/definitions/stationId"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Sensor Data Retrive login successfully",
            "schema": {
              "$ref": "#/definitions/stationId"
            }
          }
        }
      }
    },


    "/user/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "User Login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User data that we want to save",
            "schema": {
              "$ref": "#/definitions/loginUserData"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User login successfully",
            "schema": {
              "$ref": "#/definitions/loginUserData"
            }
          }
        }
      }
    },


    "/sensor/updateSensorStatus": {
      "put": {
        "tags": [
          "SensorData"
        ],
        "description": "Sensor Data Update",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Sensor data that we want to Update",
            "schema": {
              "$ref": "#/definitions/updateSensorData"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Sensor Data Update successfully",
            "schema": {
              "$ref": "#/definitions/updateSensorData"
            }
          }
        }
      }
    },



    "/weather/getAllWeatherData": {
      "get": {
        "tags": [
          "WeatherData"
        ],
        "summary": "Get All Saved Weather Data",
        "responses": {
          "200": {
            "description": "Weather Data Is Found. With Last 10 Weather Data",
            "schema": {
              "$ref": ""
            }
          }
        }
      }
    },

    "/user/getAllUsers": {
      "get": {
        "tags": [
          "UserData"
        ],
        "summary": "Get All Saved User Data",
        "responses": {
          "200": {
            "description": "User Data Is Found. With Last 10 Weather Data",
            "schema": {
              "$ref": ""
            }
          }
        }
      }
    }

  },
  "definitions": {
    "capturedWeatherData": {
      "properties": {
        "station_id": {
          "type": "integer"
        },
        "temperature": {
          "type": "string"
        },
        "humidity": {
          "type": "string"
        },
        "light_intensity": {
          "type": "string"
        },
        "rain_water_level": {
          "type": "string"
        },
        "pressure": {
          "type": "string"
        },
        "air_quality": {
          "type": "string"
        },
        "noise": {
          "type": "string"
        },
        "gas_detection": {
          "type": "string"
        },
        "uv": {
          "type": "string"
        },
        "rain_detection": {
          "type": "string"
        },
        "rainfall_intensity": {
          "type": "string"
        }
      }
    },

    "userData": {
      "properties": {
        "userName": {
          "type": "string"
        },
        "userAddress": {
          "type": "string"
        },
        "userTelephone": {
          "type": "string"
        },
        "userStation": {
          "type": "string"
        },
        "userEmail": {
          "type": "string"
        },
        "userPassword": {
          "type": "string"
        }
      }
    },


    "sensorData": {
      "properties": {
        "stationId": {
          "type": "string"
        },
        "sensorId": {
          "type": "string"
        },
        "sensorName": {
          "type": "string"
        },
        "activeStatus": {
          "type": "string"
        }
      }
    },


    "loginUserData": {
      "properties": {
        "userEmail": {
          "type": "string"
        },
        "userPassword": {
          "type": "string"
        }
      }
    },

    "stationId": {
      "properties": {
        "stationId": {
          "type": "string"
        }
      }
    },

    "updateSensorData": {
      "properties": {
        "stationId": {
          "type": "string"
        },
        "sensorId": {
          "type": "string"
        },
        "activeStatus": {
          "type": "string"
        }
      }
    }


  }
}